Criando um objeto fora da classe IRPF e adequando o método getImpostoPorFaixa
